:hammer: move validation logic
negative quality check to factory and legendary validation to validation
class
