+++ Java/src/main/java/com/gildedrose/item_helpers/ItemHandler.java
@@ -2,31 +2,14 @@ package com.gildedrose.item_helpers;
 
 import com.gildedrose.main.Item;
 
-import static com.gildedrose.items.LegendaryItem.NOT_LEGENDARY_ITEM_ERROR_MESSAGE;
-import static com.gildedrose.items.LegendaryItem.isNotLegendary;
 import static java.lang.Math.max;
 
 public class ItemHandler {
-
-    public static final String QUALITY_ERROR_MESSAGE = "Quality cannot be negative! Current value: ";
 
     private final Item item;
 
     public ItemHandler(Item item) {
         this.item = item;
-    }
-
-    static void validate(Item item) {
-        if (qualityIsNegative(item)) {
-            throw new IllegalArgumentException(QUALITY_ERROR_MESSAGE + item.quality);
-        }
-        if (isNotLegendary(item)) {
-            throw new IllegalArgumentException(NOT_LEGENDARY_ITEM_ERROR_MESSAGE + item.quality);
-        }
-    }
-
-    private static boolean qualityIsNegative(Item item) {
-        return item.quality < 0;
     }
 
     public boolean qualityIsHigherThanZero() {
